Issues fixed:
* Channelwise PCA accepts x only

* Local path def removed, init for folders added.

* Demo files have their main()

* Removed duplicate ToggleNumpy()

Persisting Issues:

* Redundant filter.py

* Redundant trigger_decoder.py

+++ utils/toggleNumpy.m
@@ -25,7 +25,7 @@ if isnumeric(inArray)
         
     % Flip original shape because numpy's reshape reads inputs in opposite
     % order
-    outArray = outArray.reshape(fliplr(origSize));
+    outArray = outArray.reshape(fliplr(int16(origSize)));
     return
 end
